[#51] Refactor: bootstrap wiring, integrationtests module, remove legacy source
Bootstrap:
  ContactFormMailApplication — @SpringBootApplication + @ConfigurationPropertiesScan
    (auto-discovers CaptchaProperties record in infrastructure module)
  ApplicationBeanConfig — explicit @Bean factory that wires ContactFormService
    with its three outbound port adapters; no Spring annotation in application layer

Web:
  ContactFormExceptionHandler — @RestControllerAdvice maps CaptchaVerificationException
    to HTTP 400 (replaces ResponseStatusException from original code)

application.yml — moved to contact-form-mail-bootstrap/src/main/resources
Test application.yml — copied to bootstrap and integrationtests test resources

Containerfiles updated:
  prod: multi-stage COPY by module; uses *-exec.jar from bootstrap target
  dev:  points to contact-form-mail-bootstrap/target/*-exec.jar

Deleted: src/ (legacy monolithic source tree fully replaced by modules above)
