+++ contact-form-mail/contact-form-mail-bootstrap/src/main/java/de/bot/pilot/mail/ContactFormMailApplication.java
@@ -1,0 +1,14 @@
+package de.bot.pilot.mail;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+
+@SpringBootApplication
+@ConfigurationPropertiesScan
+public class ContactFormMailApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ContactFormMailApplication.class, args);
+    }
+}
