+++ contact-form-mail/contact-form-mail-bootstrap/src/test/java/de/bot/pilot/mail/ContactFormMailApplicationTests.java
@@ -1,0 +1,12 @@
+package de.bot.pilot.mail;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ContactFormMailApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+}
