+++ contact-form-mail/contact-form-mail-integrationtests/src/test/java/de/bot/pilot/mail/ContactFormMailIT.java
@@ -1,0 +1,12 @@
+package de.bot.pilot.mail;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest(classes = ContactFormMailApplication.class)
+class ContactFormMailIT {
+
+    @Test
+    void contextLoads() {
+    }
+}
