--- contact-form-mail/src/main/java/de/bot/pilot/mail/contact/dto/CaptchaResponse.java
@@ -1,4 +1,0 @@
-package de.bot.pilot.mail.contact.dto;
-
-public record CaptchaResponse(boolean success) {
-}
