--- contact-form-mail/src/main/java/de/bot/pilot/mail/contact/persistence/repository/api/EmailRequestRepository.java
@@ -1,8 +1,0 @@
-package de.bot.pilot.mail.contact.persistence.repository.api;
-
-import de.bot.pilot.mail.contact.persistence.domain.EmailRequest;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface EmailRequestRepository extends JpaRepository<EmailRequest, Long> {
-
-}
