--- contact-form-mail/src/main/java/de/bot/pilot/mail/contact/service/api/CaptchaService.java
@@ -1,7 +1,0 @@
-package de.bot.pilot.mail.contact.service.api;
-
-import java.io.IOException;
-
-public interface CaptchaService {
-    void verify(String token, String ip) throws IOException;
-}
