--- contact-form-mail/src/test/java/de/bot/pilot/mail/ContactFormMailApplicationTests.java
@@ -1,13 +1,0 @@
-package de.bot.pilot.mail;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ContactFormMailApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
